+++ packages/graphql/src/tsp-index.ts
@@ -1,6 +1,6 @@
 import type { DecoratorImplementations } from "@typespec/compiler";
 import { NAMESPACE } from "./lib.js";
-import { $compose, $Interface } from "./lib/interface.js";
+import { $compose, $interface } from "./lib/interface.js";
 import { $operationFields } from "./lib/operation-fields.js";
 import { $mutation, $query, $subscription } from "./lib/operation-kind.js";
 import { $schema } from "./lib/schema.js";
@@ -9,7 +9,7 @@ import { $specifiedBy } from "./lib/specified-by.js";
 export const $decorators: DecoratorImplementations = {
   [NAMESPACE]: {
     compose: $compose,
-    Interface: $Interface,
+    interface: $interface,
     mutation: $mutation,
     query: $query,
     operationFields: $operationFields,